refactor: move interlacer to dicom module, __init__ add all the modalities to use it
