+++ src/imgtools/modalities/__init__.py
@@ -3,3 +3,14 @@ from .pet import PET, PETImageType
 from .scan import Scan
 from .segmentation import Segmentation, accepts_segmentations, map_over_labels
 from .structureset import StructureSet
+
+__all__ = [
+    "Dose",
+    "PET",
+    "PETImageType",
+    "Scan",
+    "Segmentation",
+    "accepts_segmentations",
+    "map_over_labels",
+    "StructureSet",
+]
